+++ services/common-services/fx-services/accounting-service/internal/pub/pub.go
@@ -30,13 +30,13 @@ type TransactionEvent struct {
 	TransactionID   int64                  `json:"transaction_id"`
 	TransactionType string                 `json:"transaction_type"` // deposit, withdrawal, transfer, etc
 	Status          string                 `json:"status"`
-	Amount          int64                  `json:"amount"` // Atomic units
+	Amount          float64                `json:"amount"` // Atomic units
 	Currency        string                 `json:"currency"`
 	AccountNumber   string                 `json:"account_number,omitempty"`
 	FromAccount     string                 `json:"from_account,omitempty"`
 	ToAccount       string                 `json:"to_account,omitempty"`
-	BalanceAfter    int64                  `json:"balance_after,omitempty"`
-	Fee             int64                  `json:"fee,omitempty"`
+	BalanceAfter    float64                `json:"balance_after,omitempty"`
+	Fee             float64                `json:"fee,omitempty"`
 	ErrorMessage    string                 `json:"error_message,omitempty"`
 	Metadata        map[string]interface{} `json:"metadata,omitempty"`
 	Timestamp       time.Time              `json:"timestamp"`
@@ -51,18 +51,18 @@ func (p *TransactionEventPublisher) PublishTransactionEvent(ctx context.Context,
 		return fmt.Errorf("failed to marshal event: %w", err)
 	}
 
-	if err := p.rdb. Publish(ctx, TransactionEventsChannel, payload).Err(); err != nil {
-		return fmt. Errorf("failed to publish event: %w", err)
+	if err := p.rdb.Publish(ctx, TransactionEventsChannel, payload).Err(); err != nil {
+		return fmt.Errorf("failed to publish event: %w", err)
 	}
 
 	log.Printf("[TransactionEvent] Published: %s for user=%s, receipt=%s",
-		event.EventType, event.UserID, event. ReceiptCode)
+		event.EventType, event.UserID, event.ReceiptCode)
 
 	return nil
 }
 
 // PublishTransactionCompleted publishes a successful transaction completion
-func (p *TransactionEventPublisher) PublishTransactionCompleted(ctx context.Context, userID, receiptCode string, transactionID int64, transactionType, currency string, amount, balanceAfter, fee int64) error {
+func (p *TransactionEventPublisher) PublishTransactionCompleted(ctx context.Context, userID, receiptCode string, transactionID int64, transactionType, currency string, amount, balanceAfter, fee float64) error {
 	return p.PublishTransactionEvent(ctx, &TransactionEvent{
 		EventType:       "transaction.completed",
 		UserID:          userID,
@@ -78,7 +78,7 @@ func (p *TransactionEventPublisher) PublishTransactionCompleted(ctx context.Cont
 }
 
 // PublishTransactionFailed publishes a failed transaction
-func (p *TransactionEventPublisher) PublishTransactionFailed(ctx context.Context, userID, receiptCode string, transactionID int64, transactionType, currency string, amount int64, errorMsg string) error {
+func (p *TransactionEventPublisher) PublishTransactionFailed(ctx context.Context, userID, receiptCode string, transactionID int64, transactionType, currency string, amount float64, errorMsg string) error {
 	return p.PublishTransactionEvent(ctx, &TransactionEvent{
 		EventType:       "transaction.failed",
 		UserID:          userID,
@@ -93,7 +93,7 @@ func (p *TransactionEventPublisher) PublishTransactionFailed(ctx context.Context
 }
 
 // PublishDepositCompleted publishes a deposit completion
-func (p *TransactionEventPublisher) PublishDepositCompleted(ctx context.Context, userID, receiptCode, accountNumber string, amount int64, currency string, balanceAfter int64) error {
+func (p *TransactionEventPublisher) PublishDepositCompleted(ctx context.Context, userID, receiptCode, accountNumber string, amount float64, currency string, balanceAfter float64) error {
 	return p.PublishTransactionEvent(ctx, &TransactionEvent{
 		EventType:       "deposit.completed",
 		UserID:          userID,
@@ -108,7 +108,7 @@ func (p *TransactionEventPublisher) PublishDepositCompleted(ctx context.Context,
 }
 
 // PublishWithdrawalCompleted publishes a withdrawal completion
-func (p *TransactionEventPublisher) PublishWithdrawalCompleted(ctx context.Context, userID, receiptCode, accountNumber string, amount int64, currency string, balanceAfter int64) error {
+func (p *TransactionEventPublisher) PublishWithdrawalCompleted(ctx context.Context, userID, receiptCode, accountNumber string, amount float64, currency string, balanceAfter float64) error {
 	return p.PublishTransactionEvent(ctx, &TransactionEvent{
 		EventType:       "withdrawal.completed",
 		UserID:          userID,
@@ -123,7 +123,7 @@ func (p *TransactionEventPublisher) PublishWithdrawalCompleted(ctx context.Conte
 }
 
 // PublishTransferCompleted publishes a transfer completion
-func (p *TransactionEventPublisher) PublishTransferCompleted(ctx context.Context, userID, receiptCode, fromAccount, toAccount string, amount int64, currency string, fee int64) error {
+func (p *TransactionEventPublisher) PublishTransferCompleted(ctx context.Context, userID, receiptCode, fromAccount, toAccount string, amount float64, currency string, fee float64) error {
 	return p.PublishTransactionEvent(ctx, &TransactionEvent{
 		EventType:       "transfer.completed",
 		UserID:          userID,
